+++ src/lib.rs
@@ -29,7 +29,7 @@ mod session;
 mod sndlink;
 
 pub use self::connection::{Connection, ConnectionController};
-pub use self::errors::AmqpTransportError;
+pub use self::errors::{AmqpTransportError, LinkError};
 pub use self::rcvlink::{ReceiverLink, ReceiverLinkBuilder};
 pub use self::session::Session;
 pub use self::sndlink::{SenderLink, SenderLinkBuilder};