+++ src/rcvlink.rs
@@ -97,9 +97,11 @@ impl ReceiverLink {
     }
 
     pub(crate) fn remote_closed(&self, error: Option<Error>) {
+        trace!("Receiver link has been closed remotely");
         let inner = self.inner.get_mut();
         inner.closed = true;
         inner.error = error;
+        inner.reader_task.wake();
     }
 }